[Trainer user] Move model
