+++ app/models/model_trainer_user.py
@@ -1,7 +1,10 @@
 from typing import List
 
 from flask_restx import fields, Model
-from pydantic import BaseModel, Field, field_validator
+from pydantic import BaseModel, Field
+
+from app import TrainerUser, Trainer, Users
+from app.common.constants import const
 
 user_of_trainer = Model('UserOfTrainer', {
     'user_id': fields.Integer(readOnly=True, description='The user unique identifier'),
@@ -30,8 +33,122 @@ class UserTrainer(BaseModel):
     center_location: str = Field(description="center_location")
 
 
+class CreateTrainerUserRelationRequest:
+    user_name: str
+    phone_number: str
+    lesson_total_count: int
+    lesson_current_count: int
+    exercise_days: str
+    special_notice: str
+
+    def __init__(self, data):
+        self.user_name = data['user_name']
+        self.phone_number = data['phone_number']
+        self.lesson_total_count = data['lesson_total_count']
+        self.lesson_current_count = data['lesson_current_count']
+        self.exercise_days = data['exercise_days']
+        self.special_notice = data['special_notice']
+
+
+class UpdateTrainerUserRequest:
+    lesson_total_count: int
+    lesson_current_count: int
+    exercise_days: str
+    special_notice: str
+
+    def __init__(self, data):
+        self.lesson_total_count = data['lesson_total_count']
+        self.lesson_current_count = data['lesson_current_count']
+        self.exercise_days = data['exercise_days']
+        self.special_notice = data['special_notice']
+
+
 class UserTrainersResponse(BaseModel):
     results: List[UserTrainer]
+
+
+class UsersRelatedTrainerResponse:
+    user_id: int
+    user_name: str
+    user_profile_img_url: str
+    exercise_days: str
+    lesson_current_count: int
+    lesson_total_count: int
+    registered_date: str
+    last_date: str
+
+    @staticmethod
+    def to_dict(trainer_user: TrainerUser, user: Users):
+        return {
+            "user_id": trainer_user.user_id,
+            "user_name": user.user_name,
+            "user_profile_img_url": user.user_profile_img_url,
+            "exercise_days": trainer_user.exercise_days,
+            "lesson_current_count": trainer_user.lesson_current_count,
+            "lesson_total_count": trainer_user.lesson_total_count,
+            "registered_date": trainer_user.created_at.strftime(const.DATETIMEFORMAT),
+            "last_date": trainer_user.deleted_at.strftime(
+                const.DATETIMEFORMAT) if trainer_user.deleted_at is not None else None
+        }
+
+
+class TrainersRelatedUserResponse:
+    trainer_id: int
+    trainer_name: str
+    trainer_profile_img_url: str
+    lesson_name: str
+    lesson_total_count: int
+    lesson_current_count: int
+    center_name: str
+    center_location: str
+
+    @staticmethod
+    def to_dict(trainer_user: TrainerUser, trainer: Trainer):
+        return {
+            "trainer_id": trainer.trainer_id,
+            "trainer_name": trainer.trainer_name,
+            "trainer_profile_img_url": trainer.trainer_profile_img_url,
+            "lesson_name": trainer.lesson_name,
+            "lesson_total_count": trainer_user.lesson_total_count,
+            "lesson_current_count": trainer_user.lesson_current_count,
+            "center_name": "Center 아직 적용 안함",
+            "center_location": "Center 아직 적용 안함"
+        }
+
+
+class UserDetailRelatedTrainerResponse:
+    name: str
+    email: str
+    gender: str
+    phone_number: str
+    user_profile_img_url: str
+    delete_flag: bool
+    birthday: str
+    lesson_total_count: int
+    lesson_current_count: int
+    exercise_days: str
+    special_notice: str
+    registered_date: str
+    last_date: str
+
+    @staticmethod
+    def to_dict(user: Users, trainer_user: TrainerUser):
+        return {
+            "name": user.user_name,
+            "email": user.user_email,
+            "gender": user.user_gender,
+            "phone_number": user.user_phone_number,
+            "user_profile_img_url": user.user_profile_img_url,
+            "delete_flag": user.user_delete_flag,
+            "birthday": user.user_birthday,
+            "lesson_total_count": trainer_user.lesson_total_count,
+            "lesson_current_count": trainer_user.lesson_current_count,
+            "exercise_days": trainer_user.exercise_days,
+            "special_notice": trainer_user.special_notes,
+            "registered_date": trainer_user.created_at.strftime(const.DATETIMEFORMAT),
+            "last_date": trainer_user.deleted_at.strftime(
+                const.DATETIMEFORMAT) if trainer_user.deleted_at is not None else None
+        }
 
 
 user_detail_of_trainer = Model('User Details of Trainer', {
