+++ app/services/service_trainer_user.py
@@ -5,7 +5,7 @@ from app.entities.entity_trainer_user import TrainerUser
 from app.entities.entity_users import Users
 from app.repositories.repository_trainer import TrainerRepository
 from app.repositories.repository_users import UserRepository
-from app.routes.models.model_trainer_user import UsersRelatedTrainerResponse, CreateTrainerUserRelationRequest, \
+from app.models.model_trainer_user import UsersRelatedTrainerResponse, CreateTrainerUserRelationRequest, \
     TrainersRelatedUserResponse, UserDetailRelatedTrainerResponse, UpdateTrainerUserRequest
 from app.services.service_image import ImageService
